+++ changelog.md
@@ -5,6 +5,7 @@
 * Add `recras-input-invalid` class to invalid amount inputs
 * Use minimum amount instead of number of people for "minimum amount" calculation
 * Fix book button being enabled despite "I agree" checkbox not being checked
+* Show line price based on amount selected
 
 ## 0.10.3 (2018-11-20)
 * Fix standalone contact forms